s3 bucket: renamed s3bucketpolicy client/controller to bucketpolicy
