+++ pkg/clients/s3/bucketpolicy.go
