+++ pkg/clients/s3/fake/bucketpolicy.go
