+++ pkg/controller/s3/bucketpolicy/bucketpolicy_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package s3bucketpolicy
+package bucketpolicy
 
 import (
 	"context"
